code/requirements.txt
redis
flask
mysql-connector-python